Add inventory to the game; Handle looting on server side.
Add inventory completely into the game, by creating inventory and inventoryItem objects
on server and client side. Client side still lacking add/remove functions for inventory,
but for now we fully update the inventory with each change.

Handle loot on server side - loot was handled on server side, but the client handled
it how he wanted on his side - picked up loot even though the inventory was full or
regardless of the rules defined in the server. Making the server handle all non-UI
stuff is probably the task that will follow this project for a long time, one
action at a time :)