Add skill bar to the game; Add spells to the game.
Added a skill bar in the middle of the bottom bar. Also introduced spells to the game - a SkillSlot in a Skillbar may hold either a Spell or a (usable) Item.
Added SkillSlot - takes a keyBind and a Spell/Item and will .use them if that key is pressed.
Added Spell - takes a 'kind' argument. Later on cooldown, cast time, tooltip, name, etc. will be loaded from a database.
Added Skillbar - .add(keyBind, Spell/Item) - if there's room on the skillbar, creates a SkillSlot with the given key and Spell/Item.
		 .remove(idx) - removes SkillSlot at index 'idx'.
		 .click(key) - searches for a SkillSlot binded to key 'key' and .use's it.