Fix player enter series of events; Simplify health and damage event broadcasting
PlayerEnter event was executed wrongly due to lack of callback usage, leading to not placing the player in the correct "map group" when he enters the game.
This lead to having to change zones in order to get the current zone spawns.

Health and damage events were split upon few events, leading to confusion about which should be sent on what occasion, or even double notifications.
Now there are two events for this - Health (entityId, hp, maxHP [, isRegen]) and Damage (receiverId, damage, attackerId). The latter is simply for combat logs and
notifications purposes, while the former actually updates hp of entities.
Also note that now these events are being broadcasted rather than sent to a specific player.