Added more chat channels; Added Party system; Added sprintf lib
Added Party system to the game, usable from text commands - /invite <player>, /accept <player>, /leave, /leader <player>, /kick player.
Added Party and Yell channels to the chat system, while also showing indication of which channel the chat is set to and beautifully presenting the chat log.

As part of the Party and Chat system, there was a need to keep track of players regardless of spawned entities in order to interact with them while not in range.
Fixed huge player spawn/respawn bugs.