Fixed auto equipping of armor/weapon
Previously, we would equip the armor/weapon on pick up only on the client side and not actually from the server/inventory.
Now, after a successful loot event given from the server, we will equip the item through the correct inventoryitem and be in sync with the server.
NOTE: We will still optimistically auto equip on the client, to give the effect of a fast game, but in practice we are verifying with the server first.